requirements.txt
pyTelegramBotAPI==4.14.0
anthropic==0.49.0
python-dotenv==1.0.0
openai>=1.66.2
openai-agents==0.0.4
chromadb==0.4.22
pydantic>=2.10.0,<3
griffe>=1.5.6,<2
typing-extensions>=4.12.2,<5
requests>=2.0.0,<3
spacy>=3.6.0,<4
pillow>=10.0.0,<11
langchain>=0.1.0,<0.2.0
langchain-community>=0.0.10,<0.1.0
faiss-cpu>=1.7.4,<2
tiktoken>=0.5.1,<0.6.0
# Download the English model with: python -m spacy download en_core_web_md 
telebot==0.0.4
pytelegrambotapi==4.7.0
python-dotenv==0.21.0
anthropic==0.9.0
pydantic==1.10.2
pytest==7.2.0
pytest-mock==3.10.0
pytest-asyncio==0.20.3
pytest-cov==4.0.0
openai==1.8.0
tenacity==8.2.2
asyncio==3.4.3
aiohttp==3.8.4
requests==2.28.1
beautifulsoup4==4.12.2
Pillow==10.0.0 